Removing styles from Send, adding InputRecipient component
