+++ src/components/dashboard/Send.js
@@ -1,14 +1,11 @@
 import React, { useEffect, useState } from 'react'
-import { StyleSheet, View } from 'react-native'
-import { Icon, normalize } from 'react-native-elements'
-import { HelperText, TextInput } from 'react-native-paper'
 import isEmail from 'validator/lib/isEmail'
 import { BackButton, useScreenState } from '../appNavigation/stackNavigation'
 import userStorage from '../../lib/gundb/UserStorage'
 import logger from '../../lib/logger/pino-logger'
 import { readCode } from '../../lib/share'
 import { useDialog } from '../../lib/undux/utils/dialog'
-import Clipboard from '../../lib/utils/Clipboard'
+import InputRecipient from '../../components/common/InputRecipient'
 import isMobilePhone from '../../lib/validators/isMobilePhone'
 import goodWallet from '../../lib/wallet/GoodWallet'
 import { CustomButton, IconButton, Section, TopBar, Wrapper } from '../common'
@@ -107,39 +104,18 @@ const Send = props => {
     return response
   }
 
-  const pasteToWho = async () => {
-    try {
-      const who = await Clipboard.getString()
-      log.info({ who })
-      setScreenState({ to: who })
-    } catch (err) {
-      log.error('Paste action failed', err)
-    }
-  }
-
   return (
     <Wrapper>
       <TopBar push={props.screenProps.push} />
-      <Section style={styles.bottomSection}>
-        <Section.Stack grow={1} justifyContent="space-evenly">
-          <Section.Title style={styles.title}>TO WHOM?</Section.Title>
-          <View style={styles.iconInputContainer}>
-            <View style={styles.pasteIcon}>
-              <Icon size={normalize(16)} color="#282c34" name="content-paste" onPress={pasteToWho} />
-            </View>
-            <TextInput
-              onChangeText={text => setScreenState({ to: text })}
-              onBlur={checkError}
-              value={to}
-              error={error}
-              style={styles.input}
-              placeholder="Phone Number / Email / Username"
-              autoFocus
-            />
-          </View>
-          <HelperText type="error" visible={error}>
-            {error}
-          </HelperText>
+      <Section grow>
+        <Section.Title>Send To?</Section.Title>
+        <Section.Stack grow={1} justifyContent="space-around">
+          <InputRecipient
+            onChangeText={text => setScreenState({ to: text })}
+            onBlur={checkError}
+            to={to}
+            error={error}
+          />
         </Section.Stack>
         <Section.Row grow={2} justifyContent="flex-end">
           <ScanQRButton screenProps={props.screenProps} disabled={!!to} />
@@ -161,45 +137,5 @@ const Send = props => {
 Send.navigationOptions = {
   title: SEND_TITLE
 }
-
-const styles = StyleSheet.create({
-  title: {
-    fontSize: normalize(24),
-    fontWeight: '500'
-  },
-  centered: {
-    justifyContent: 'center',
-    alignItems: 'baseline'
-  },
-  bottomSection: {
-    flex: 1,
-    paddingTop: normalize(22),
-    justifyContent: 'space-between'
-  },
-  bottomContainer: {
-    width: '100%',
-    flexDirection: 'row',
-    marginTop: '1rem'
-  },
-  topContainer: {
-    flex: 1
-  },
-  iconInputContainer: {
-    display: 'inline-flex',
-    position: 'relative'
-  },
-  input: {
-    flex: 1,
-    backgroundColor: 'inherit',
-    marginTop: normalize(10)
-  },
-  pasteIcon: {
-    position: 'absolute',
-    cursor: 'pointer',
-    right: 0,
-    paddingTop: normalize(30),
-    zIndex: 1
-  }
-})
 
 export default Send
